requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
firebase-admin==6.2.0
gunicorn==21.2.0
django-ratelimit==4.1.0
python-dotenv==1.0.0
requests==2.31.0
